kernel: reorganize platform module
