+++ kernel/src/process.rs
@@ -10,7 +10,7 @@ use crate::capabilities;
 use crate::errorcode::ErrorCode;
 use crate::ipc;
 use crate::mem::{ReadOnlyAppSlice, ReadWriteAppSlice};
-use crate::platform::mpu::{self};
+use crate::traits::mpu::{self};
 use crate::sched::Kernel;
 use crate::syscall::{self, Syscall, SyscallReturn};
 use crate::upcall::UpcallId;
