+++ kernel/src/traits/watchdog.rs
